refactor(webui): remove dead ios native av1 experiment gate and fix codec tests
