+++ apps/webui/src/features/player/utils/engineExperiment.ts
@@ -1,10 +1,10 @@
-import { isTruthyOverrideValue } from './iosNativeAv1Experiment';
+import { isTruthyOverrideValue } from './overrideFlags';
 
 // Runtime opt-in for routing Safari through hls.js + ManagedMediaSource instead
 // of native HLS. Stage 0 ships this flag but DOES NOT consume it yet — the
 // engine selection is untouched, so behaviour is identical with the flag on or
-// off until the Stage 1 wiring lands. Default OFF, no hostname auto-enable
-// (unlike iosNativeAv1Experiment): this must be an explicit, per-device A/B.
+// off until the Stage 1 wiring lands. Default OFF, no hostname auto-enable:
+// this must be an explicit, per-device A/B.
 //
 // Enable:  ?xg2g_hlsjs_safari=1  or  localStorage XG2G_HLSJS_SAFARI=1
 // Kill:    localStorage XG2G_HLSJS_SAFARI_KILL=1  (checked first, forces OFF)
